backend/requirements.txt
python-dotenv>=1.1.0

langchain-core>=0.2.38
langchain-community>=0.2.0
langchain-openai>=0.1.0

faiss-cpu>=1.8.0
rank-bm25==0.2.2

httpx>=0.28.1

fastapi>=0.111.0
pydantic>=2.7.0
uvicorn>=0.30.0
sqlalchemy>=2.0.30